requirements.txt
llama-index-core
llama-index-llms-ollama
llama-index-embeddings-ollama
llama-index-readers-file

# slackbot
slackclient
slackeventsapi
python-dotenv
flask
